proper git tracking of refactor
